requirements.txt
blinker==1.5
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
et-xmlfile==1.1.0
Flask==2.2.2
Flask-Dance==6.2.0
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
numpy==1.23.4
oauthlib==3.2.2
openai==0.24.0
openpyxl==3.0.10
pandas==1.5.1
pandas-stubs==1.5.1.221024
psycopg2==2.9.5
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.5
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
SQLAlchemy==1.4.42
SQLAlchemy-Utils==0.38.3
tqdm==4.64.1
types-pytz==2022.5.0.0
typing_extensions==4.4.0
urllib3==1.26.12
URLObject==2.4.3
Werkzeug==2.2.2
